+++ src/Components/SparkHub/Results/Response/index.module.scss
