+++ src/Components/SparkHub/index.jsx
@@ -1,7 +1,7 @@
 import { useSelector } from "react-redux";
-import Operations from "../Operations";
-import OverlayMessage from "../OverlayMessage";
-import Results from "../Results";
+import Operations from "./Operations";
+import OverlayMessage from "./OverlayMessage";
+import Results from "./Results";
 import classes from "./index.module.scss";
 
 const SparkHub = () => {
